// src/Type/validation/signUpValType.type.ts
export type signUpValType = {
    valSignInUserNameRequir: string;
    ValSignInUserNameMin: string;
    ValSignInUserNameMax: string;
    valContactUsEmailRequir: string;
    ValContactUsEmailNotValid: string;
    valSignInPasswordRequir: string;
    ValSignInPasswordMin: string;
    ValSignInPasswordMax: string;
    valSignInConfirmPasswordRequir: string;
    ValSignInConfirmPasswordMin: string;
    ValSignInConfirmPasswordMax: string;
    ValSignInConfirmPasswordNotMatch: string;
    usernameExistError: string;
    EmailExistError: string;
};